requirements.txt
discord.py>=2.3.0
wavelink>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
